added standard templates fro register/login. modified gig class to reduce db queries (3=>1) for create method. changed genre from int to byte for consistency. using mvc form helpers in create view

+++ GigHub/ViewModels/GigFormViewModel.cs
@@ -9,7 +9,7 @@ namespace GigHub.ViewModels
         public string Time { get; set; }
 
         //set type as int b/c we're getting the genre id
-        public int Genre { get; set; }
+        public byte Genre { get; set; }
 
         public IEnumerable Genres { get; set; }
     }
